refactor(services): extract web session lifecycle and seat clients
Ten write-side methods move out of the composition root: createSession,
deleteSession, switchSession, renameSession, pinSession, unpinSession,
archiveSession and unarchiveSession into web-session-lifecycle-client.ts,
and updateSessionSeats plus rebindRemoteSessionSshConnection into
web-session-seat-client.ts. Split in two because one module holding all
ten would have crossed the 300-line rule.

deleteSession is the method that used to fuse chat, recovery and chat
configs into one component. It moves whole and reaches all three through
their state modules' accessors, which is what the chat cut was for.

The file-local tr(key) shim goes away with its last two callers; the
moved methods call i18n.t directly, which is what tr forwarded to and
what the other extracted modules already use.

web-agent-client.ts 1105 -> 822, agent-service.ts 320 -> 308.
Subtree aggregate 19319 -> 19347.

tsc --noEmit, contracts:check and the full suite pass: 286 files /
1301 tests, unchanged, with no test file edited.
